[ refactor ] Hero Component 삭제 및 LinkCreator, ShareDescription 재배치
